Weak configure-module validation
The user_domain could be missing since the action is now called by
restore-module and clone-module and at that time the user domain could
be still not available.

Removing the validation of user_domain allows to set a reference to a
non-existing domain that hopefully is soon to appear in the cluster.