fixup! Fix Postfix relayrules table
Change the primary key definition for the relayrules table.

A rule_subject can appear in one rule at most, actually it becomes the
primary key. It seems it's unlikely to happen that in the real world the
same subject is used for different rule_type. Furthermore the NS7
implementation was based on the same assumption.

Refs

- commit daa16c03630aa0404a380233ce49dad67b5d4c35
- pr #95