fix: limit LDAP queries to certain domains (#191)
If no domain with addusers flag exists, ensure LDAP query based on the
domain is skipped. We define a special "notempty.invalid" domain name to
ensure that the domain filter is always non-empty.

The same workaround is implemented for the addgroups flag. Ensure LDAP
queries based on the domain are skipped if no domain has addgroups flag.

Refs NethServer/dev#7385